requirements.txt
# Discord and bot framework
discord.py>=2.3.2
wavelink>=3.2.0

# Environment and configuration
python-dotenv>=1.0.0
PyNaCl>=1.4.0

# External integrations
spotipy>=2.22.1
lyricsgenius>=3.0.1

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Web dashboard (optional)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
aiofiles>=23.2.1

# HTTP requests
requests>=2.31.0
aiohttp>=3.8.0

# YouTube support
yt-dlp>=2023.11.16

# Utilities
colorlog>=6.8.0
psutil>=5.9.6
Pillow>=10.1.0
